siberemare-multiagent-v2/requirements.txt
langgraph>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-neo4j>=0.3.0
pydantic>=2.10.0
pillow>=11.0.0
pytesseract>=0.3.13
slack-sdk>=3.33.0
jira>=3.2.0
neo4j>=5.25.0
click>=8.1.7
fpdf2>=2.7.9
python-dotenv>=1.0.1
structlog>=24.4.0
asyncpg>=0.29.0
redis>=5.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
langchain-groq>=0.2.0
matplotlib>=3.9.0
aiohttp>=3.9.0
rich>=13.7.0
